refactor: directory reform
# Model, View 폴더 생성

- Develop 브랜치에 합칠 때 충돌 가능성 줄이기 위함
