--- TexTris/src/main/java/org/Stech/SE5/config/data/config.json
@@ -1,1 +1,0 @@
-{"GameMode":"BASIC","GameDifficulty":"NORMAL","BoardSize":"MEDIUM","BoardWidth":10,"BoardHeight":20,"GameSpeed":1.0,"ColorBlindMode":false,"KeyBindingLength":7,"moveBlockLeft":37,"moveBlockRight":39,"moveBlockDown":40,"clockwiseRotate":38,"dropBlockAtOnce":32}
