+++ TexTris/src/main/java/org/Stech/SE5/config/maintenance/SetConfig.java
@@ -1,10 +1,10 @@
-package org.Stech.SE5.config;
+package org.Stech.SE5.config.maintenance;
 
 import java.io.BufferedWriter;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.Stech.SE5.config.KeySetting.keyBinding;
+import static org.Stech.SE5.config.maintenance.KeySetting.keyBinding;
 
 public class SetConfig {
     public static void saveConfig() {
