--- TexTris/src/main/java/org/data/Setting.json
@@ -1,34 +1,0 @@
-{
-  "displayMode": 0,
-  "displayList": [
-    {
-      "x": 0,
-      "y": 0,
-      "width": 1280,
-      "height": 720
-    },
-    {
-      "x": 0,
-      "y": 0,
-      "width": 1120,
-      "height": 630
-    },
-    {
-      "x": 0,
-      "y": 0,
-      "width": 960,
-      "height": 540
-    }
-  ],
-  "isColorBlind": false,
-  "moveBlockLeft": 37,
-  "moveBlockRight": 39,
-  "moveBlockDown": 40,
-  "clockwiseRotate": 38,
-  "dropBlockAtOnce": 32,
-  "moveLeft2Key": 65,
-  "moveRight2Key": 68,
-  "moveDown2Key": 83,
-  "rotate2Key": 87,
-  "stack2Key": 82
-}
